Refactor Github admin
